Enhance DOM manipulation.
Keep track of the initial array size and checks if the current size is larger or smaller to help nodes management.
Hoist all nodes to manipulate them if there's a need to add or remove.
Added a minor update to the model logic to prevent calling the reactive callback for iterations. (causing infinite loops).

#14